Refactored Plaid::Utils to Plaid::Connection
