+++ lib/plaid.rb
@@ -1,6 +1,5 @@
 require 'plaid/version'
 require 'plaid/config'
-require 'plaid/util'
 
 require 'plaid/add_user'
 require 'plaid/user/user'
@@ -8,12 +7,11 @@ require 'plaid/institution/institution'
 require 'plaid/category/category'
 
 module Plaid
+  autoload :Connection, 'plaid/connection'
+
   class << self
     # Configures the gem with the public, private, and environment vars
     include Plaid::Configure
-
-    # Include the utility classes used throughout the gem
-    include Plaid::Util
 
     # Includes the method to authenticate the user. Defined in auth.rb
     include Plaid::AddUser
@@ -35,14 +33,14 @@ module Plaid
     # Builds an institution object and returns when the institution details exist
     def institution(id = nil)
       _institution = Plaid::Institution.new
-      res = self.get('institutions',id)
+      res = Plaid::Connection.get('institutions',id)
       id.nil? ? _institution.instantiate_all_institutions(res) : _institution.instantiate_one_institution(res)
     end
 
     # Builds an institution object and returns when the category details exist
     def category(id = nil)
       _category = Plaid::Category.new
-      res = self.get('categories',id)
+      res = Plaid::Connection.get('categories',id)
       id.nil? ? _category.instantiate_all_categories(res) : _category.instantiate_one_category(res)
     end
